// src/store/notes/types.ts
export interface Note extends ApiResponse {
  id: number;
  title: string;
}

export type ApiResponse = Record<string, any>;

export enum NotesActionTypes {
  FETCH_ALL_REQUEST = 'NOTES/FETCH_ALL_REQUEST',
  FETCH_ALL_SUCCESS = 'NOTES/FETCH_ALL_SUCCESS',
  FETCH_ALL_ERROR = 'NOTES/FETCH_ALL_ERROR',
  FETCH_DETAIL_REQUEST = 'NOTES/FETCH_DETAIL_REQUEST',
  FETCH_DETAIL_SUCCESS = 'NOTES/FETCH_DETAIL_SUCCESS',
  FETCH_DETAIL_ERROR = 'NOTES/FETCH_DETAIL_ERROR',
  ADD_REQUEST = 'NOTES/ADD_REQUEST',
  ADD_SUCCESS = 'NOTES/ADD_SUCCESS',
  ADD_ERROR = 'NOTES/ADD_ERROR',
  EDIT_REQUEST = 'NOTES/EDIT_REQUEST',
  EDIT_SUCCESS = 'NOTES/EDIT_SUCCESS',
  EDIT_ERROR = 'NOTES/EDIT_ERROR',
  DELETE_REQUEST = 'NOTES/DELETE_REQUEST',
  DELETE_SUCCESS = 'NOTES/DELETE_SUCCESS',
  DELETE_ERROR = 'NOTES/DELETE_ERROR',
  DISMISS_STATUS = 'NOTES/DISMISS_STATUS',
}

export enum NotesStatusTypes {
  FETCH_ALL_START = 'FETCH_ALL_START',
  FETCH_ALL_DONE = 'FETCH_ALL_DONE',
  FETCH_DETAIL_START = 'FETCH_DETAIL_START',
  FETCH_DETAIL_DONE = 'FETCH_DETAIL_DONE',
  ADD_START = 'ADD_START',
  ADD_DONE = 'ADD_DONE',
  EDIT_START = 'EDIT_START',
  EDIT_DONE = 'EDIT_DONE',
  DELETE_START = 'DELETE_START',
  DELETE_DONE = 'DELETE_DONE',
}

export interface NotesState {
  status: NotesStatusTypes | undefined;
  data: Note[];
  note: Note | undefined;
  errors?: string;
}